+++ my-campaign/src/components/get_home/get_home.css
@@ -1,0 +1,15 @@
+.header-get-home {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    padding: 0px;
+    gap: 20px;
+
+    position: absolute;
+    width: 1380px;
+    height: 64px;
+    left: 0px;
+    top: 0px;
+
+    background: #FFFFFF;
+}
