+++ my-campaign/src/components/logo/logo.css
@@ -1,5 +1,5 @@
 
-.container {
+.container-logo {
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -14,7 +14,7 @@
     margin-left: 40px;
   }
   
-  .container__image {
+  .container-logo__image {
     width: 100%;
     height: 100%;
   }
@@ -35,11 +35,11 @@
     color: #FFF8F8;
   }
   
-  .container__text {
+  .container-logo__text {
     width: 56px;
     height: 24px;
   
-    font-family: 'Raleway';
+    font-family: 'Edwardian Script ITC', Courier, monospace;
     font-style: normal;
     font-weight: 400;
     font-size: 10px;
