+++ my-campaign/src/pages/create_followers_page/CreateFollowersPage.jsx
@@ -1,5 +1,5 @@
 import Footer from '../../components/footer/Footer';
-import GetHome from '../../components/GetHome';
+import GetHome from '../../components/get_home/GetHome';
 import CreateFollowerPage from '../../components/CreateFollowers';
 
 function CreateFollower() {
