+++ my-campaign/src/pages/create_users_page/CreateUserPage.jsx
@@ -1,5 +1,5 @@
 import Footer from '../../components/footer/Footer';
-import GetHome from '../../components/GetHome';
+import GetHome from '../../components/get_home/GetHome';
 import CreateUsers from '../../components/CreateUsers';
 
 function CreateUserPage() {
