+++ my-campaign/src/styles.css
@@ -16,16 +16,125 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
-.title{
+
+.title {
   font-size: 40px;
   font-weight: bold;
+  font: #4A0243;
 }
 
-.subtitle{
+.subtitle {
   font-size: 30px;
-  font-weight:bolder;
+  font-weight: bolder;
 }
-.paragraph{
+
+.paragraph {
   font-size: 16px;
   font-weight: lighter;
+}
+
+.form {
+  height: 654px;
+  width: 534px;
+  left: 423px;
+  top: 86px;
+  border-radius: 8px;
+  padding: 40px;
+
+  box-sizing: border-box;
+
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 40px;
+  gap: 8px;
+
+  position: absolute;
+  left: 30.65%;
+  right: 30.65%;
+  top: 10.21%;
+  bottom: 12.01%;
+
+  background: #FFFFFF;
+  border: 2px solid #4A0243;
+  border-radius: 8px;
+}
+
+.input-text {
+  width: 100%px;
+  height: 40px;
+  left: 0px;
+  top: 26px;
+  border-radius: 8px;
+  padding: 8px 16px 8px 16px;
+
+  box-sizing: border-box;
+
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  padding: 8px 16px;
+  gap: 8px;
+
+  width: 434px;
+  height: 40px;
+
+  background: #FFFFFF;
+  border: 1px solid #000000;
+  border-radius: 8px;
+}
+
+.label {
+  width: 100%px;
+  height: 22px;
+
+}
+
+.row {
+  width: 100%;
+  height: 66px;
+  left: 40px;
+  top: 112px;
+  border-radius: 0px;
+  padding: 0px 10px 0px 10px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding: 0px 10px;
+  gap: 4px;
+
+  width: 454px;
+  height: 66px;
+}
+
+.button {
+  width: 454px;
+  height: 44px;
+  color: white;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  padding: 10px 1px 10px 10px;
+  gap: 10px;
+
+
+  background: #4A0243;
+  border-radius: 8px;
+}
+
+.main {
+  position: relative;
+  width: 1380px;
+  height: 842px;
+
+  background: url('./images/fondo.jpg');
+
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 20px;
+
 }
